+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - PR #2380 New cudf::is_sorted checks whether cudf::table is sorted
 - PR #2356 Java column vector standard deviation support
 - PR #2415 Revamp `value_counts` to use groupby count series of any type
+- PR #2437 ORC reader: Add 'use_np_dtypes' option
 - PR #2396 Add support for `cudf.util.hash_cudf_object` & refactor `cudf.utils` to `cudf.util`
 
 ## Improvements